Fix a subtle `StringWriter` bug which can manifest when the memory block allocated to the destination string is adjacent to the memory block of the internal buffer of the `StringWriter`.
In essence, the `start() == &dest[0]` condition can be `true` if `start()` is
the end of a memory block unrelated to the payload of `dest`.

To solve this, avoid such comparisons. Use `secondary_buffer_.empty()` instead,
which is slightly different, because it covers also some cases where
`start_to_limit() == 0`, so code guarded by that condition must not assume that
`start() == &dest[0]`, removing minor optimizations.

PiperOrigin-RevId: 442806025