Add an alternative `std_io` design which replaces singletons of standard streams (`Std{In,Out,Err}()`) with the possibility to create a `Reader` or `Writer` for a standard stream on demand (`NewStd{In,Out,Err}`).
For output streams, this introduces natural and explicit flushing points at
`Close()`, analogously to other writers, instead of calling `StdOut().Flush()`
(which is unusual) or relying on a `std::atexit()` handler (which does not
guarantee proper ordering wrt. other handlers).

This simplifies rules of interoperability between `Std{Out,Err}` and other means
of accessing standard streams. Instead of flushing the `Writer` and leaving it
open but inactive, which is brittle, it can be closed and later recreated.

This allows to tune their options, e.g. buffering.

Finally, this encourages parameterizing code by a `Reader` or `Writer` or their
factory, which makes code easier to reuse and test than if the code accessed a
singleton directly.

A cost of the change is that it is no longer possible to inject other types of
streams to the global state. Only file descriptors can be injected, which is
meant for testing of code which is not parameterized by a `Reader` or `Writer`.

Usages of old functions will be migrated to new classes. Then old functions will
be removed and new classes will be renamed by dropping the `New` prefix.

PiperOrigin-RevId: 442847129