Cosmetics: order `BUILD` rules for readers before the corresponding writers, instead of the other way around.
The original ordering was so because writers were deemed more fundamental than
readers (`Reader` depends on `Writer` but `Writer` did not depend on `Reader`
back then), one has to write before there is anything to read, and that ordering
is more natural for compressors/decompressors and encoders/decoders.

But that ordering feels awkward in other contexts, because common conventions
are reading/writing and input/output. These contexts seem to be more common.

Hence the switch. Only `chunk_encoding` rules stay unchanged because they are
named encoder/decoder and compressor/decompressor.

PiperOrigin-RevId: 445329226