Cosmetic changes to compressor code.
If getting Zstd uncompressed size failed during construction because the source
was too short, but the source could have grown in the meantime, try again.

Update comments: it is OK to supply a dictionary for `ZstdReader` if no
dictionary was supplied for `ZstdWriter`, like in `ZlibReader`. This is not true
in case of `BrotliReader`.

Fix whitespace in assertions.

Reorder members of `ZstdReaderBase` to more closely match the order of usage.

PiperOrigin-RevId: 446477221