In `StringWriter`, reduce the capacity of a smallish `std::string` (affecting sizes between 23 and 255 bytes on Linux with modern libc++ and with default `StringWriterBase::Options`).
This is done by lowering the minimal capacity of the first non-SSO block from
255 bytes to the current value of `dest.size() * 1.5` (some bound besides
`max(min_length, recommended_length)` is used to ensure more than 1 byte of
growth even if `std::string::reserve()` sets the minimal possible capacity).

The cost of this change is doing more allocations if the string turns out to be
larger but the first `Push()` after SSO is done with a small `min_length` and
`recommended_length`.

PiperOrigin-RevId: 446699218