Fix a bug in `Limiting{,Backward}Writer` when the limit is exceeded: do not move the cursor in `FailLimitExceeded()`, because it is called also when the buffer is already reset. Move it in `SyncBuffer()` instead.
Also, fix comment about `SyncBuffer()` precondition which was no longer true.

(Strictly speaking, moving the cursor might be skipped altogether, because
currently it is unspecified how much is written if the limit is exceeded.
But `Limiting{,Backward}Writer` might be changed to guarantee writing everything
up to the limit, in which case this will become useful.)

PiperOrigin-RevId: 447500918