Do not bother shifting `{ReaderInput,WriterOutput}Stream::ByteCount()` from the initial `{Reader,Writer}::pos()` to 0. It is less surprising if `ByteCount()` simply reflects `pos()`.
Even though the documentation of `ZeroCopy{Input,Output}Stream::ByteCount()`
says “total number of bytes read/written since this object was created”, generic
code deals only with already existing objects, and thus it cannot tell whether
the position increased before or after constructing the stream adapter. The
absolute meaning of positions is important only for code which creates an object
of a concrete class, and that code can in principle reason whether the position
is correct, or even use `PrefixLimiting{Reader,Writer}` to guarantee shiftting
the initial position to 0.

Reading or writing past the range of `int64_t` continues to be treated like
failure, so that `ByteCount()` is always consistent with reading/writing.

PiperOrigin-RevId: 447717240