Fix setting `ZlibReaderBase::Options::set_window_log()`. It was ignored.
Change the type of `window_log` in `ZlibReader` from `absl::optional<int>` to
`int`.

The difference between passing an explicit default (internally expressed as 0)
and the maximum (15) was not very useful, so `ZlibReader` might as well default
to 15. For zlib headers 0 allocates the size implied by the header, while 15
allocates 32k, but either choice limits the maximum accepted distance to what
the header specifies. For gzip headers 0 is interpreted as 15; `window_log` is
not stored in that header. In raw mode there is no way to pass 0 because raw
mode is requested by negating `window_log`.

This allows to use `ZlibReaderBase::Header::kRaw` without specifying
`window_log` explicitly.

PiperOrigin-RevId: 449010612