Let `SnappyWriter` fail when trying to write 4GB of data or more.
The Snappy format does not support larger sizes (the compressed stream begins
with a varint32-encoded uncompressed size), and the native compressor does not
check whether the size is in the supported range (when given more data, it
produces a compressed stream which does not decompress to the original).

The implementation is a bit tricky because it supports `Push()` with a
`min_length` which allows `limit_pos()` to overflow
`std::numeric_limits<uint32_t>::max()`, as long as it does not overflow
`std::numeric_limits<size_t>::max()`. If `cursor()` moves too far then the
overflow is detected in `SyncBuffer()`.

PiperOrigin-RevId: 449682494