Improve handling of position overflow during reading: even if reading at least a particular length is requested, read as much as remains until the maximum position and then fail, instead of failing before reading anything.
This does not matter much in practice, because such positions are normally not
reached, but an upcoming change will sometimes ask an internal function to read
more data than the public caller asked for. It does not feel right to
unnecessarily fail immediately.

This also makes handling a known exact size more regular: reading past that size
is never attempted, no matter whether the size is reached between reads or in
the middle of a read.

PiperOrigin-RevId: 449683946