Fix `{Zstd,Lz4}Reader` uncompressed position overflow detection (which was temporarily disabled in the case of `ZstdReader`).
The previous code could mistakenly believe that the position overflows in
certain rare circumstances:

* The whole contents are being decompressed to a flat array with
  `ZSTD_d_stableOutBuffer`, otherwise the loop would have exited as soon as
  enough data have been produced.

* The whole output buffer has been filled, i.e. all data have been decompressed
  and the buffer did not have space for more, but not the whole input was read
  yet (some footer remains), so `ZSTD_decompressStream()` did not signal end of
  stream yet.

In this case `output.pos == output.size` for a different reason than output.size
having been reduced to `std::numeric_limits<Position>::max() - limit_pos()`.
The core of the faulty reasoning was that it is possible that all output has
been produced but decompression is not finished yet.

To fix this, check for `input.pos < input.size` after `ZSTD_decompressStream()`
returns and after other exit conditions are checked. This is possible only if
the output buffer is full, while less than `min_length` was output, so either
the output buffer size has been reduced due to overflow, or
`ZSTD_decompressStream()` wants to output more than the expected decompressed
size to a flat buffer. The conditions expected to hold are asserted.

Note that directly checking
`output.pos == std::numeric_limits<Position>::max() - limit_pos()` would
technically be too strict, because this space might be barely enough.

In contrast to reading from a file, calling the decompressor with no output
space is correct, because it might be finishing reading the input without
producing more output. Hence the overflow is checked after the decompressor
returns, rather than before calling it.

The case of `Lz4Reader` is analogous to `ZstdReader`.

PiperOrigin-RevId: 449997910