Change `ZlibReader` position overflow detection to be analogous to `{Zstd,Lz4}Reader`.
In contrast to `{Zstd,Lz4}Reader`, the old way was correct for `ZlibReader`,
because `ZlibReader` does not have an optimized mode which forces the loop to go
on filling a flat buffer until the end of the stream. `ZlibReader` returns when
at least `min_length` was output, which can be less than the buffer size only
when the position overflows.

This is being changed for consistency.

PiperOrigin-RevId: 450400690