Do not increase buffer lengths to `max_buffer_size` when `SetReadAllHint(true)` is used but the source size is not known.
This is a tradeoff and the previous resolution was too risky.

For small sources, this reduces memory usage (which can be important), and
longer reads are not useful in that case anyway. For very large sources, the
increase of the number of reads is relatively small. Only medium-sized sources
might see a meaningfully larger number of reads, but the difference is a
constant number of extra reads: `log2(max_buffer_size) - log2(min_buffer_size)`.

PiperOrigin-RevId: 451820799