Fix `SerializeToWriter()` when the `Writer` failed during final flushing of the buffer of the `CodedOutputStream`: report the failure instead of possibly leaving the `Writer` as failed but reporting success.
The fix is to call `CodedOutputStream::Trim()` before checking `dest.ok()`
(alternatively the `CodedOutputStream` could have been destroyed, but then it
would be inaccessible for an assertion later).

If `Trim()` is not called, then the destructor of `CodedOutputStream` flushes
its buffer, and that might cause the `Writer` to fail, but this is too late to
report because `dest.ok()` was checked earlier.

Luckily this was not a big problem in practice, because the `Writer` remains
failed, so the failure will be reported by `Close()`. Except that
`coded_stream.HadError()` used in the assertion also flushes the buffer, and
thus the assertion might have failed in debug mode.

PiperOrigin-RevId: 454782827