Reduce moves and allocations in `Dependency<Ptr, AnyDependency<Ptr, _>&&>`, in particular if `AnyDependencyRef<Ptr>` is initialized with `AnyDependency<Ptr, _>&&` (when it cannot have internals adopted).
In these cases there is no need to decay `AnyDependency<Ptr, _>&&` to
`AnyDependency<Ptr, _>`, which requires moving it to the heap since it does not
fit in inline storage of `AnyDependencyRef<Ptr>`. It is more efficient to refer
to the `AnyDependency<Ptr, _>` staying in its original location.

This is done by allowing specializations of `DependencyImpl<Ptr, M&&>` (besides
`DependencyImpl<P*, M&&>` with `M*` convertible to `P*`), instead of always
decaying them to `DependencyImpl<Ptr, std::decay_t<M>>`. They are still decayed
if they do not have dedicated specializations.

To avoid ambiguous specializations, this in turn is done by inserting an
intermediate specialization layer between `DependencyImpl` and `Dependency`,
called `dependency_internal::DependencyMaybeRef`, which handles cases when
`Manager` is a reference: by delegating to a specialization of `DependencyImpl`
if it exists, otherwise by decaying the reference to a value type.

PiperOrigin-RevId: 455623045