Optimize setting up the initial buffer when `SetWriteSizeHint()` is used:
In `StringWriter` and `{Chain,Cord}{,Backward}Writer`, do not set buffer
pointers during initialization. If `SetWriteSizeHint()` is later used, the
initial buffer likely points to inline storage which is too small, and its
contents might need to be copied, so it is better to delay setting buffer
pointers to another operation, which is more likely to happen after
`SetWriteSizeHint()`.

To compensate for this, `Cord{,Backward}Writer` need to be able to decide to use
the short buffer in `PushSlow()`. `StringWriter` and `Chain{,Backward}Writer`
were already able to do this.

Simplify setting the initial buffer pointers of `Cord{,Backward}Writer`:
`size_hint_` is never known during initialization nowadays since it is no longer
specified in `Options`.

PiperOrigin-RevId: 457962120