Add `ResizableWriter`.
It generalizes `StringWriter` to other objects with a flat representation, e.g.
`std::string` with a non-default allocator.

The implementation strategy is slightly different because some resizables might
not track their allocated capacity separately from the size, and resizing
downwards and back upwards would trigger unnecessary reallocations. Hence
`ResizableWriter` delays resizing the destination until required, and even if
data are written to the secondary buffer, the destination might keep free space
appended.

Another reason for differences is the templated nature of the destination type.
Hence combinations of operations specific to the destination type are grouped in
virtual functions of `ResizableWriterBase`, so that the rest of the code of
`ResizableWriterBase` can be independent from the destination type.

PiperOrigin-RevId: 459453465