Add `AssertInitialized()` function to assert that a region of memory is initialized, which is checked when running under memory sanitizer.
Use it to verify that memory is actually written under
`{,Backward}Writer::cursor()` before calling functions which may embed that
memory in the destination.

This makes detection of violations more consistent, also if the particular
operation in the particular `Writer` does not yet move the cursor and thus gives
another chance to fill the memory, and also when the destination is ultimately
not observed.

PiperOrigin-RevId: 459526321