Assert that memory being written with `{,Backward}Writer::Write()` is initialized.
This complements the existing checks of buffer contents by detecting writing of
uninitialized data closer to the point when they are being written, instead of
only after the current buffer overflows (or never if they happen to be embedded
in a `Chain` or `Cord`). This makes it easier to see where the data are being
written, not just where they were allocated.

PiperOrigin-RevId: 460220359