Remove `LimitingReader` and `Limiting{,Backward}Writer` preconditions that the current position must not already exceed the limit.
If it does, proceed in the same way as if it reached the limit later: pretend
that the source ends (for `LimitingReader` by default) or fail (for
`LimitingReader` with the `exact()` option, or for `Limiting{,Backward}Writer`).
In any case the current position is moved back to not exceed the limit.

PiperOrigin-RevId: 462350799