Optimize constructing a `CsvHeader` from an iterable by caching the last payload created, and reusing it the next time if it matches.
Reusing `CsvHeader` directly is more efficient but not always feasible.

This asked for reworking how constructors and `{,Try}Reset()` are redirected to
common code. The public functions now have a fewer number of overloads
(`std::vector<std::string>&&` is handled by the generic version), and they all
redirect to `TryResetInternal()` which performs further dispatching depending on
how the cache can be examined. For consistency the same has been done for
`CsvRecord`.

PiperOrigin-RevId: 470004402