Unify `Writer::SupportsSize()` with `SupportsRandomAccess()`.
Remove `Writer::Size()` implementations where `!SupportsRandomAccess()`, i.e.
mostly those which trivially return `pos()` or delegate to another `Writer`, and
also `riegeli::tensorflow::FileWriter::Size()`.

Reasons:

1. `Size()` as different from `pos()` is useful only when we can `Seek()`.

2. The only reasonable way to detect random access support in `FILE` and
   `std::ostream` is to try to seek to the end, which breaks if they
   successfully seek to the end but do not support random access in general.
   This was the case of `Writer{CFile,OStream}` if they wrap a `Writer` which
   supports size but not random access.

   It would be possible for `Writer{CFile,OStream}` to refuse to tell the size
   if the `Writer` supports `Size()` but not `Seek()`, in order to fix random
   access autodetection, but it is simpler to continue supporting the concept of
   `Size()` at all only when `Seek()` is supported too.

In `WriterCFile`, reflect `ReadMode()` only if `SupportsRandomAccess()`,
removing the special case of allowing to seek if the `Reader` but not the
`Writer` supports that, so that it is possible to detect whether the resulting
`FILE` supports random access by trying to seek to the end.

`WriterOStream` is not changed like `WriterCFile` because
`std::streambuf::seekoff()` distinguishes between seeking for reading and
writing.

State that if `Writer::SupportsRandomAccess()`, then any `Reader` obtained from
`ReadMode()` must `Reader::SupportsRandomAccess()`. This allows to replace
failure paths which are never meant to happen with assertions.

PiperOrigin-RevId: 474994289