In `{Fd,CFile}Writer`, do not claim supporting random access in append mode, because writing might always happen at the end.
In `CFileWriter` detecting the append mode of an already opened `FILE` is in
general impossible, so indicating the append mode in
`CFileWriterBase::Options::mode()` for an already open file got more important,
and `OStreamWriterBase::Options::assumed_append()` got added.

Also in `Fd{Reader,Writer}`, if `assumed_pos()` and `independent_pos()` are both
set, fail gracefully instead of treating this as a precondition violation
detected only in debug mode.

PiperOrigin-RevId: 474995203