Check for random access support in `{Fd,CFile,IStream}Reader` eagerly instead of lazily.
This is because most of the time (unless `growing_source()`) this has a side
effect of getting to know `exact_size()`, which affects buffer sizing. While
delaying the check might avoid syscalls, it leads to worse buffer sizing
decisions.

Random access was often checked early anyway, in particular in `Pull()` and
`Read()` through `ToleratesReadingAhead()` if `!read_all_hint()`. So the
difference is observable only if `SetReadAllHint(true)` was used.

Checking for random access support is superfluous only if `growing_source()`,
but since this is not the default, for simplicity laziness is not left merely
for that case.

PiperOrigin-RevId: 474995568