If an `AnyDependency` gets moved to an `AnyDependency` with different template arguments, decide whether to adopt it (instead of wrapping) at runtime if this decision cannot be made at compile time.
Even if the destination has a smaller statically known inline size or alignment,
the actual object in the source might be small enough to fit. Or the source
might keep the object behind the pointer, which is always safe to adopt.

Also, announce that `GetIf()` is unspecified in this case. It was already true.

PiperOrigin-RevId: 475762334