Change `Dependency<Ptr, std::reference_wrapper<M>>::manager()` to return `std::reference_wrapper<M>&` rather than `M&`, and generally handle `std::reference_wrapper` regularly as a `Manager` type rather than as a mode to pass a `Manager`. `AnyDependency::GetIf()` no longer treats it specially.
This is necessary because lots of code (in particular `src()` and `dest()`
members of `Reader` and `Writer` classes) assume that
`Dependency<Ptr, Manager>::manager()` returns exactly `Manager&`, which breaks
for `std::reference_wrapper`. Introducing some abstraction for that code to know
which type to return would make source code more complicated, so it is simpler
to treat `std::reference_wrapper` like a regular `Manager` type.

The cost of this change is that it is no longer allowed to wrap an arbitrary
argument specifying a dependent object in `std::ref()` just in case, e.g. when
passing a raw pointer or `std::unique_ptr`. `std::ref` is applicable solely to
an owned object of the right type or to `AnyDependency`.

This requires extra constraints in `AnyDependencyImpl` constructors (which
exclude the case when the argument is `std::reference_wrapper` applied to the
same `AnyDependencyImpl` specialization) for a very subtle reason. Luckily it is
not needed to understand the reason to accept that the constraints are at least
harmless.

The reason is that declaring a copy constructor of `std::reference_wrapper<T>`
has a surprising side effect of instantiating templates needed for a
hypothetical conversion from `const std::reference_wrapper<T>&` to `T`
(which is normally not defined and does not need to). Checking
`IsValidDependency<Ptr, std::reference_wrapper<AnyDependencyImpl>>` instantiates
`DependencyBase<std::reference_wrapper<AnyDependencyImpl>>`, which needs a copy
constructor of `std::reference_wrapper<AnyDependencyImpl>`, which instantiates
templates needed for a hypothetical conversion from
`const std::reference_wrapper<AnyDependencyImpl>&` to `AnyDependencyImpl`, which
checks `IsValidDependency<Ptr, std::reference_wrapper<AnyDependencyImpl>>`,
which is already in the process of being determined, so this fails the build
with a bad usage of an incomplete type.

PiperOrigin-RevId: 475798948