Improve `AnyDependency<void*>` support.
If `P` is possibly cv-qualified `void`, then `Dependency<P*, Manager>`
has an ambiguous interpretation for `Manager` being `M*`, `nullptr_t`,
`std::unique_ptr<M, Deleter>`, or `std::reference_wrapper<M>`. The ambiguity
is resolved in favor of pointing the `void*` to the dereferenced `M`, not to
the pointer object itself.

PiperOrigin-RevId: 481122902