Use `::tensorflow::Status()` instead of conditionally `::tensorflow::OkStatus()` or `::tensorflow::Status::OK()` depending on the TF version.
The version with `#if TF_GRAPH_DEF_VERSION >= 1137` was in fact not working because the header containing it is not included, so it was always defaulting to the `::tensorflow::Status::OK()` branch.

I replaced many occurrences with `Status()`, it's standard, and `OkStatus() { return Status(); }`, so I don't think there is any issue related to idiomatic or not idiomatic here.

PiperOrigin-RevId: 484466053