In `AnyDependency` inlining criterion, check if the `Dependency` is move constructible rather than if `Manager` is move constructible, because the `Dependency` is what is being moved. This is usually the same because most `Dependency` instances have a single member of type `Mananager`, but in principle not necessarily.
Make `Dependency` constructors from `const Manager&` and `Manager&&` optional
with SFINAE, instead of claiming that they are always supported yet failing if
the `Manager` type cannot support them. This was needed for fixing the inlining
criterion in `AnyDependency`.

Add SFINAE guards to `AnyDependency{,Ref}` methods parameterized by `Manager`.
In particular `GetIf()` may be called only with a type for which there is a
possibility of a match.

Remove exclusion of `std::reference_wrapper<AnyDependency{,Ref}>` in
`AnyDependency{,Ref}` constructor and assignment. This used to not work but
works thanks to the following.

Tweak `IsValidDependency`. For subtle reasons explained in comments it should
sometimes avoid instantiating the corresponding `DependencyImpl`. One of special
cases was needed for adding SFINAE to `DependencyBase` constructors, the other
was needed for removing the exclusion of `std::reference_wrapper`.

PiperOrigin-RevId: 486089246