Mark certain types with `ABSL_ATTRIBUTE_TRIVIAL_ABI` (if supported).
This introduces a more efficient calling convention (the object is passed by
registers or by value on the stack rather than by address), at the cost of a
different lifetime and ABI (the object passed by value is destroyed by the
callee at the end of the call rather than by the caller at the end of the full
expression).

If a type has trivial copy/move constructors and destructor, trivial ABI is
already used implicitly. Otherwise `ABSL_ATTRIBUTE_TRIVIAL_ABI` is applicable if
an object of the given type is sometimes passed by value, and if it can be
safely moved to a different address as if by `memcpy()` and skipping destroying
the original.

`ABSL_ATTRIBUTE_TRIVIAL_ABI` is effective only if base classes and non-static
member variables have trivial ABI.

The following types are marked:
* `AnyDependency{,Ref}` when `Ptr` has trivial ABI and `inline_size == 0`
  (if the stored `Dependency` has non-trivial ABI then it will be forced to be
  non-inlined)
* `ObjectState`
* `RefCountedPtr`
* `Buffer`
* `OwnedFd`
* `OwnedCFile` (conditional on `std::unique_ptr` having trivial ABI)
* `FutureRecordPosition`

The following types do not need to be marked because they have trivial ABI
implicitly:
* `SharedBuffer`
* `ChainBlock`
* `{Zstd,Lz4,Zlib}Dictionary` (`ZlibDictionary` is conditional on
  `std::shared_ptr` having trivial ABI; `BrotliDictionary` would require
  `std::vector` to have trivial ABI)
* `CsvHeader` (`CsvRecord` would require `std::vector` to have trivial ABI)

Further explanation:
https://quuxplusone.github.io/blog/2018/05/02/trivial-abi-101/.

PiperOrigin-RevId: 486134286