Add `Text{Reader,Writer}` for newline conversion, primarily between LF and the native representation.
It will be used for working with Windows text files without relying on Windows
text mode in I/O. This is applicable also for cases when file contents are not
written directly to a file, e.g. under compression.

It was easier to implement them with the newline representation known
statically, because `kLf` can avoid buffering. Hence variants with runtime
choice are separate as `AnyText{Reader,Writer}`.

PiperOrigin-RevId: 491555790