Add `ClosingPtr()` (returning `std::unique_ptr<T, NullDeleter>`) to replace `std::ref()` (returning `std::reference_wrapper<T>`) in the context of a `Dependency` or `AnyDependency`.
It does not need a separate specialization of `DependencyImpl`: rules for
`std::unique_ptr` already cover it with the intended semantics (the only
difference is that `Release()` releases the pointer instead of returning
`nullptr`). This reduces cognitive load: it works like a `std::unique_ptr` which
nominally owns but does not destroy the object because it is exactly that.

It does not trip on a subtlety which required
`IsValidDependencyImpl<Ptr, std::reference_wrapper<AnyDependencyImpl<Ptr>>>`
to be explicitly specialized.

Uses of `std::ref()` in this context will be migrated to `ClosingPtr()`.

PiperOrigin-RevId: 493867798