Fix and simplify free space counting in a block, which might have resulted in unnecessary large allocations:
* Remove `space_{after,before}()` which implicitly adjusted the space to
  `capacity()` if `empty()`. Rename `raw_space_{after,before}()` to
  `space_{after,before}()` instead, i.e. remove the adjustment there.

* Move the adjustment to their usages: `can_{append,prepend}()` and
  `Can{Append,Prepend}MovingData()`.

This fixes the adjustment in `CanAppendMovingData()` which could have been
performed twice in opposite directions. If `empty()` then `space_after()` is
adjusted to `capacity()` (this is good, empty contents can cheaply moved to the
left when appending) but `space_before_if_not` was adjusted to `capacity()` too
(this is not the correct amount of space before contents which should be
maintained, because contents is not being moved to te right). Dually in
`CanPrependMovingData()`.

PiperOrigin-RevId: 498009653