For moving contents of a block or the array of block pointers instead of reallocating, it is enough when the existing array has enough capacity and is at currently most half full. Do not require that the new size it at most half the capacity. This switches to moving when existing contents are small but at the wrong end and the new size fills the whole array.
Ensure amortized growth of block capacity only if the block is mutable at the
time of being enlarged. If it cannot be mutated, the scenario asks for a new
block no matter what was the previous block, so its size was not important.

When moving or reallocating contents of a block or the array of block pointers,
leave space at both ends only when space is needed at the front. When space is
needed at the back, do not leave any space at the front.

This reduces moves or reallocations when elements are removed from the front and
added at the back. This also simplifies code a bit. At the same time this is
enough to maintain a constant amortized cost of allocations for any pattern of
adding and removing elements, even though sometimes two moves or reallocations
are needed when elements are added at the back and at the front.

PiperOrigin-RevId: 498695963