Replace `ChainBlock` with `SizedSharedBuffer`.
Differences:

* `SizedSharedBuffer` allocates the data using `SharedBuffer`, i.e. `Buffer`,
  i.e. with a standalone allocation without a header.

  This might reduce memory allocator waste (depending on the memory allocator)
  in cases when the buffer capacity is a power of 2 which will be more common.

* External objects, i.e. custom deleters, are not supported.

* Options are not supported. In particular there is no size hint, and the
  minimal size is whatever is requested instead of 256 bytes.

`Chain::BlockIterator::Pin()` returns a new type `Chain::PinnedBlock` instead of
`ChainBlock`. `PinnedBlock` is a simple wrapper over `RefCountedPtr<RawBlock>`.

Add `SharedBuffer::ToCord() &&` overload and
`SharedBuffer::{Append,Prepend}SubstrTo()`. They are used by
`SizedSharedBuffer`. Also, change their implementation to embed the
`SharedBuffer` as the external object of the `Cord` instead of using `Share()`
and `DeleteShared()`.

Strengthen the precondition of
`{,Shared}Buffer::{ToCord,{Append,Prepend}SubstrTo}()`: for the substring to not
be contained in the buffer, require not only `length == 0` but also
`data == nullptr`. This is consistent with `SizedSharedBuffer::Substr()`, which
must match its invariants because `SizedSharedBuffer::Substr() &&` modifies the
object in place (to avoid losing the allocation).

Rename `SharedBuffer:const_data()` to `data()`.

Apply `std::move()` to `buffer_` when possible (i.e. when the buffer is then
cleared anyway) in `BufferedReader` to avoid redundant refcount operations.

In `Pushable{,Backward}Writer`, restore buffer allocation also after the buffer
is written as a `Chain`. It is possible that the conversion from
`SizedSharedBuffer&&` to `Chain` will not drop the allocation after all (if the
buffer is wasteful).

PiperOrigin-RevId: 502593622