Reduce the allocation of the buffer or scratch if they are getting wasteful.
This not only reduces memory usage after a singular event which requires a large
buffer, but this also avoids a particular inefficient behavior. If the buffer or
scratch once got very large (e.g. because of `Pull(min_length)`), then during
normal operations it was never shared when reading to `Chain` or `absl::Cord`
because that would be wasteful, hence data were always copied from it, getting
stuck in the state of having a large allocation.

Let `{,Shared}Buffer::Reset()` automatically drop the allocation if it would be
wasteful. Also, handle the parameter of 0 to avoid allocating. This lets e.g.
`BufferedWriter` reduce the buffer size.

Add `SizedSharedBuffer::{,ClearAnd}Shrink()` to drop the allocation if it would
be wasteful for the given size.

Reduce refcount operations in `PullableReader` by applying `std::move()` to
`SizedSharedBuffer` if the whole remaining part of the scratch is getting
appended to the destination. This forced extracting a part of `SyncScratch()`
which does not verify preconditions and invariants, because they no longer hold
when the scratch is physically released.

Avoid unnecessary reallocation of the `SharedBuffer::Payload` object in
`SharedBuffer::Reset()` if `has_unique_owner()` but only the capacity is
insufficient.

PiperOrigin-RevId: 502595046