Let `{Read,Write}BufferSizer` align positions to multiples of sufficiently large powers of 2, and let `exact_size()` trump `min_length` in `ReadBufferSizer::BufferLength()`.
Aligned positions might make I/O and compression/decompression more efficient.

This also yields a different strategy for filling the remaining part of the
buffer if the source returned less data than asked for. When deciding whether to
fill the remaining part or to make a new buffer, the old approach was to always
fill whenever there was non-empty space. The new approach is to request the
recommended length, and fill only if that much was available; this now works
because the recommended length decreases if the previous buffer was not filled.

Get rid of allocating one byte more than `exact_size()` during reading. This
used to make it easier to verify that the source ends there when `exact_size()`
was called `size_hint()` and might have been inaccurate, but with the current
semantics of `exact_size()` the `Reader` classes are not supposed to read more
from the underlying source anyway.

Handle limiting reads to `exact_size()` by `BufferedReader` and `FileReader`
instead of particular overrides of `ReadInternal()`. This is important to avoid
allocating another buffer at the end of a file, and this removes special
handling for that in `PushSlow()` in decompressors, which now set `exact_size()`
when encountering the end of data to avoid `BufferedReader` allocating another
buffer.

Add a virtual `BufferedReader::ExactSizeReached()`. Decompressors need to
override it to ensure that the compressed stream is fully consumed after
decompressing all data. This is especially important for `ZstdReader` which can
have `exact_size()` known since initialization, and it could be 0 in which case
`ReadInternal()` is never called.

Override `BufferedReader::SizeImpl()` to return `exact_size()`.

Remove separate `LengthTo{Read,Write}Directly()` functions. If `BufferLength()`
is called again at a position within the previous buffer, it usually returns the
length of the remaining part of that buffer (unless the current run started at
an odd position and the granularity of positions just increased), so this
usually became the same as `BufferLength(pos())`, i.e. `available()` is equal to
that if not 0.

Remove `ApplyReadSizeHint()`. Rename `ApplyWriteSizeHint()` to
`ApplySizeHint()`, removing its `multiple_runs` parameter.
Add `ApplyBufferConstraints()`, which is often used near `ApplySizeHint()`.
Add `ChainBlock::{Append,Prepend}BufferIfExisting()`.

PiperOrigin-RevId: 503189936