Add `Xz{Reader,Writer}` for xz / lzma compression.
In highest density settings the compression density is comparable to brotli
(smaller or larger depending on the data), while xz is much faster to compress
and much slower to decompress. Compared to zstd, xz produces smaller data but is
slightly slower to compress and much slower to decompress.

Xz is generally not recommended for new usages because of slow decompression,
but there are existing usages with this format which would benefit from a
consistent Riegeli API.

PiperOrigin-RevId: 510126191