Move the `return false` aspect of `Fail()` to an inline function.
This allows the compiler to generate slightly better code by statically knowing
the result of `Fail()`, avoiding a branch in some cases. Even if a failure path
is rarely taken, the original implementation forced the successful continuation
path to consider also coming back from `Fail()` returning `true`, which cannot
happen.

The cost is an explicit loadof a false result on `return Fail(…)`, and losing
the opportunity of a tail call to `Fail()`. The latter is impractical anyway
because the non-trivial `absl::Status` parameter is passed by value, destroyed
by the caller (it does not currently have `ABSL_ATTRIBUTE_TRIVIAL_ABI`).

PiperOrigin-RevId: 511194314