Let `CFile{Reader,Writer}` default to opening the file with `O_CLOEXEC`, with `set_inheritable()` option to specify otherwise.
This is consistent with `Fd{Reader,Writer}`.

This is specified with the 'e' modifier of `fopen()`, which supported by most
flavors of Unix and is planned for standardization
(https://www.austingroupbugs.net/view.php?id=411), or the 'N' modifier on
Windows.

MacOS X does not support 'e', although it has `O_CLOEXEC`, and it supports only
a small number of other modifiers ('+', 'b', 'x'), so in this case `fopen()` is
emulated with `open()` + `fdopen()`.

PiperOrigin-RevId: 512011753