Let `RecordWriter::FutureFlush()` return `FutureStatus` instead of `FutureBool`, with `get()` returning `absl::Status` instead of `bool`.
This fixes propagating the status in case of asynchronous failure. Previously
there was no obvious way to get that status. The comment that `Flush()` is
equivalent to `FutureFlush().get()` was wrong: it was normal for `status()` to
be `ok()` after asynchronous failure. `RecordWriter` cannot have its `status()`
changed asynchronously, as that would race with synchronous calls.

As a workaround it would have been possible to call `Flush()` when `get()`
returned `false` in order to propagate the status, but that is not obvious.
It is better to return `FutureStatus` instead, as there is no other object
whose `status()` can be queried.

PiperOrigin-RevId: 517052297