Add `RecordReaderBase&` parameter to the recovery function.
Sometimes it is useful to access the `RecordReader` there, and arragging to
capture the variable which holds the `RecordReader` in the caller ca be
problematic, especially when the recovery function is provided in options
(when the `RecordReader` is not constructed yet) or when the `RecordReader`
is later moved.

Also, add `RecordReaderBase::set_recovery()` function to change that after
construction.

This aligns `RecordReader` with `CsvReader`.

A similar change to the Python API will be done separately.

PiperOrigin-RevId: 518547457