Add `CsvReaderBase::Options::skip_empty_lines()` option.
Let `CsvWriter` and `WriteCsvRecordToString()` write a record consiting of a
single empty field with quotes, to avoid writing an empty line which might be
skipped by some readers.

A record with no fields continues to be written as an empty line, even though
the CSV format does not support empty records and it will be read as a record
with one empty field or skipped.

When quotes are needed for UTF-8 BOM or an empty field but quoting is disabled,
skip quoting instead of failing. These cases do not need quotes according to
RFC4180 and they might be read correctly depending on `CsvReader` settings.
Quotes are being added only to aid reading without `preserve_utf8_bom()` or with
`skip_empty_lines()`.

Quote UTF-8 BOM only in the first field, not in any field.

Add some comments regarding RFC4180 conformance.

PiperOrigin-RevId: 520866074