Remove `DependencyGetIfBase`.
Instead of having to derive `DependencyImpl` specializations from
`DependencyGetIfBase` (if they are not derived from `DependencyBase`),
`Dependency` automatically provides `GetIf()` methods if `DependencyImpl`
does not define them itself.

While doing so, define static `GetIf()` variants in terms of dynamic `GetIf()`
variants if only dynamic ones are provided. Make use of that in
`DependencyImpl<AnyDependency>` specializations.

Let `StableDependency` work even when `Ptr` is not a pointer. This is now
simpler to do because `StableDependency` implementations of `operator*`,
`operator->`, comparisons with `nullptr`, and `GetIf()` are provided by an
internal `DependencyDerived` class, reused from `Dependency`.

Aff SFINAE guards to `GetIf()` so that it is present only if `OtherManager` is
potentially valid for the given `Ptr`.

PiperOrigin-RevId: 522068586